# (PC-30127)[API] chore: Improve `find_offerers_validated_3_days_ago_with_no…