…_venues` performance

The cron calling this function could take up to AN HOUR to run as soon as there's recently validated offerers.

Using `not_in` is catastrophic for performance as it trigger seq scan over several huge tables

Before:

```
staging> EXPLAIN ANALYZE SELECT offerer.address AS offerer_address, offerer.street AS offerer_street, offerer.id AS offerer_id, offerer."isActive" AS "offerer_isActive", offerer."validationStatus" AS "offerer_validationStatus", offerer."postalCode" AS "offerer_postalCod
 e", offerer.city AS offerer_city, offerer."dateCreated" AS "offerer_dateCreated", offerer.name AS offerer_name, offerer.siren AS offerer_siren, offerer."dateValidated" AS "offerer_dateValidated", offerer."allowedOnAdage" AS "offerer_allowedOnAdage"
 FROM offerer
 WHERE (offerer.id NOT IN (SELECT venue."managingOffererId"
 FROM venue
 WHERE NOT venue."isVirtual")) AND (offerer.id NOT IN (SELECT venue."managingOffererId"
 FROM venue JOIN offer ON venue.id = offer."venueId"
 WHERE venue."isVirtual" AND offer."venueId" = venue.id)) AND offerer."isActive" IS true AND CAST(offerer."dateValidated" AS DATE) = '2024-05-30'
+----------------------------------------------------------------------------------------------------------------------------------------+
| QUERY PLAN                                                                                                                             |
|----------------------------------------------------------------------------------------------------------------------------------------|
| Seq Scan on offerer  (cost=27038.89..982947446733.80 rows=46 width=118) (actual time=75.358..75.364 rows=0 loops=1)                    |
|   Filter: (("isActive" IS TRUE) AND (NOT (hashed SubPlan 1)) AND (("dateValidated")::date = '2024-05-30'::date) AND (NOT (SubPlan 2))) |
|   Rows Removed by Filter: 43651                                                                                                        |
|   SubPlan 1                                                                                                                            |
|     ->  Seq Scan on venue  (cost=0.00..13168.35 rows=61825 width=8) (actual time=0.013..36.506 rows=62391 loops=1)                     |
|           Filter: (NOT "isVirtual")                                                                                                    |
|           Rows Removed by Filter: 43643                                                                                                |
|   SubPlan 2                                                                                                                            |
|     ->  Materialize  (cost=13715.98..44314909.53 rows=90170906 width=8) (never executed)                                               |
|           ->  Hash Join  (cost=13715.98..43511824.00 rows=90170906 width=8) (never executed)                                           |
|                 Hash Cond: (offer."venueId" = venue_1.id)                                                                              |
|                 ->  Seq Scan on offer  (cost=0.00..42927355.63 rows=217420763 width=8) (never executed)                                |
|                 ->  Hash  (cost=13168.35..13168.35 rows=43810 width=16) (never executed)                                               |
|                       ->  Seq Scan on venue venue_1  (cost=0.00..13168.35 rows=43810 width=16) (never executed)                        |
|                             Filter: "isVirtual"                                                                                        |
| Planning Time: 0.604 ms                                                                                                                |
| Execution Time: 75.490 ms                                                                                                              |
+----------------------------------------------------------------------------------------------------------------------------------------+
EXPLAIN 17
Time: 0.088s
```

After:
```
staging> EXPLAIN ANALYZE SELECT offerer.address AS offerer_address, offerer.street AS offerer_street, offerer.id AS offerer_id, offerer."isActive" AS "offerer_isActive", offerer."validationStatus" AS "offerer_validationStatus", offerer."postalCode" AS "offerer_postalCod
 e", offerer.city AS offerer_city, offerer."dateCreated" AS "offerer_dateCreated", offerer.name AS offerer_name, offerer.siren AS offerer_siren, offerer."dateValidated" AS "offerer_dateValidated", offerer."allowedOnAdage" AS "offerer_allowedOnAdage", EXISTS (SELECT 1
 FROM venue JOIN offer ON venue.id = offer."venueId"
 WHERE venue."managingOffererId" = offerer.id) AS "hasVenuesWithOffers", EXISTS (SELECT 1
 FROM venue
 WHERE venue."managingOffererId" = offerer.id AND NOT venue."isVirtual") AS "hasPhysicalVenues"
 FROM offerer
 WHERE offerer.id IN (SELECT offerer.id
 FROM offerer
 WHERE offerer."isActive" AND CAST(offerer."dateValidated" AS DATE) = '2024-05-30')
+--------------------------------------------------------------------------------------------------------------------------------------------------------+
| QUERY PLAN                                                                                                                                             |
|--------------------------------------------------------------------------------------------------------------------------------------------------------|
| Nested Loop  (cost=0.29..19061.91 rows=185 width=120) (actual time=0.782..11.188 rows=198 loops=1)                                                     |
|   ->  Seq Scan on offerer offerer_1  (cost=0.00..1493.26 rows=185 width=8) (actual time=0.729..9.496 rows=198 loops=1)                                 |
|         Filter: ("isActive" AND (("dateValidated")::date = '2024-05-30'::date))                                                                        |
|         Rows Removed by Filter: 43453                                                                                                                  |
|   ->  Index Scan using offerer_pkey on offerer  (cost=0.29..6.21 rows=1 width=118) (actual time=0.001..0.001 rows=1 loops=198)                         |
|         Index Cond: (id = offerer_1.id)                                                                                                                |
|   SubPlan 1                                                                                                                                            |
|     ->  Nested Loop  (cost=0.99..489848.58 rows=6175 width=0) (actual time=0.005..0.005 rows=0 loops=198)                                              |
|           ->  Index Scan using "ix_venue_managingOffererId" on venue  (cost=0.42..16.46 rows=3 width=8) (actual time=0.001..0.002 rows=2 loops=198)    |
|                 Index Cond: ("managingOffererId" = offerer.id)                                                                                         |
|           ->  Index Only Scan using "ix_offer_venueId" on offer  (cost=0.57..162414.59 rows=86278 width=8) (actual time=0.001..0.001 rows=0 loops=396) |
|                 Index Cond: ("venueId" = venue.id)                                                                                                     |
|                 Heap Fetches: 4                                                                                                                        |
|   SubPlan 2                                                                                                                                            |
|     ->  Index Scan using "ix_venue_managingOffererId" on venue venue_1  (cost=0.42..16.46 rows=2 width=0) (actual time=0.002..0.002 rows=1 loops=198)  |
|           Index Cond: ("managingOffererId" = offerer.id)                                                                                               |
|           Filter: (NOT "isVirtual")                                                                                                                    |
|           Rows Removed by Filter: 1                                                                                                                    |
| Planning Time: 1.087 ms                                                                                                                                |
| Execution Time: 11.307 ms                                                                                                                              |
+--------------------------------------------------------------------------------------------------------------------------------------------------------+
EXPLAIN 20
Time: 0.042s
```

This divides cost by 10000000 !


## But de la pull request

Ticket Jira (ou description si BSR) : https://passculture.atlassian.net/browse/PC-30127

## Vérifications

- [ ] J'ai écrit les tests nécessaires
- [ ] J'ai [relu attentivement les migrations](https://www.notion.so/passcultureapp/Clarifier-les-pratiques-de-migration-de-BDD-5f8edeba57ed4a17b80c847a74def027), en particulier pour éviter les _locks_, et je préviens les équipes Shérif et Data
- [ ] J'ai ajouté des screenshots pour d'éventuels changements graphiques